Modify type of monto and descuento.

+++ src/main/java/com/demo/sample1/api/presentation/PedidoForm.java
@@ -13,8 +13,8 @@ import lombok.Setter;
 public class PedidoForm {
     private Integer idPedido;
     private String nombre;
-    private Double monto;
-    private Double descuento;
+    private Integer monto;
+    private Integer descuento;
 
     public Pedido asPedido() {
         return Pedido.builder()
